create meat.service.js
