+++ src/services/meat.service.js
@@ -1,0 +1,88 @@
+const { toMysqlTimestampString } = require("../utils/mysql.utils");
+const {
+  createMeat,
+  findOneMeat,
+  updateMeat,
+  cancelMeat,
+} = require("../repository/meats.repostitory");
+const {
+  createStorage,
+  findOneStorage,
+} = require("../repository/storage.repostitory");
+const Meat = require("../entities/meat.entity");
+const MeatStatus = require("../enums/meatStatus.enum");
+const AttachmentType = require("../enums/attachmentType.enum");
+
+async function createMeatService(meatDTO) {
+  const savedStorageResult = await createStorage(
+    meatDTO.base64String,
+    AttachmentType.MEAT
+  );
+  const meat = new Meat(
+    null,
+    savedStorageResult.insertId,
+    meatDTO.title,
+    meatDTO.description,
+    meatDTO.maxParticipant,
+    toMysqlTimestampString(new Date(meatDTO.startTime)),
+    toMysqlTimestampString(new Date(meatDTO.endTime)),
+    MeatStatus.ONGOING,
+    toMysqlTimestampString(new Date()),
+    toMysqlTimestampString(new Date())
+  );
+  return await createMeat(meat);
+}
+
+async function updateMeatService(meatDTO) {
+  const originalMeat = await findOneMeat(meatDTO.id);
+  let newMeat = new Meat(
+    originalMeat.id,
+    originalMeat.imageStorageId,
+    meatDTO.title,
+    meatDTO.description,
+    meatDTO.maxParticipant,
+    toMysqlTimestampString(new Date(meatDTO.startTime)),
+    toMysqlTimestampString(new Date(meatDTO.endTime)),
+    originalMeat.status,
+    originalMeat.createdDate,
+    toMysqlTimestampString(new Date())
+  );
+  if (meatDTO.base64String) {
+    // only update storage if meatDTO has base64String
+    const savedStorageResult = await createStorage(
+      meatDTO.base64String,
+      AttachmentType.MEAT
+    );
+    newMeat.imageStorageId = savedStorageResult.insertId;
+  }
+  return await updateMeat(originalMeat.id, newMeat);
+}
+
+async function cancelMeatService(meatId) {
+  return await cancelMeat(meatId);
+}
+
+async function findOneMeatService(meatId) {
+  const meat = await findOneMeat(meatId);
+  const storage = await findOneStorage(meat.imageStorageId);
+  const result = {
+    id: meat.id,
+    imageUrl: storage.mediaLink,
+    title: meat.title,
+    description: meat.description,
+    maxParticipant: meat.maxParticipant,
+    startTime: meat.startTime,
+    endTime: meat.endTime,
+    status: meat.status,
+    createdDate: meat.createdDate,
+    lastModifiedDate: meat.lastModifiedDate,
+  };
+  return result;
+}
+
+module.exports = {
+  createMeatService,
+  updateMeatService,
+  cancelMeatService,
+  findOneMeatService,
+};
